refactor(index): move compare-graph-to-text + search-classification helpers to text_compare.rs
The graph-vs-text compare/classify helpers (graph_only_reason, classify_text_only_hit,
recommended_graph_text_fallback, compare_pattern_match_mode, parser-gap/false-positive
predicates) and the search doc-ranking/dedup helpers (rank_docs_for_symbol,
docs_locality_rank, dedupe_search_hits, bounded_summary, module_stem, is_test_like_path)
are query-support logic that had accreted in the index spine. Group them in
text_compare.rs (sibling of query_api, which is their sole caller alongside staleness).
pub(crate) + re-export. index/mod.rs 622 -> 417 — now the IndexDatabase definition,
status/result types, and the module index. No behavior change.
